test: extract custom matcher for validation to file
