+++ src/interpreter/map-validator-advanced.test.ts
@@ -5,160 +5,11 @@ import {
 } from '@superfaceai/ast';
 
 import { parseMap, parseProfile, Source } from '..';
-import { ProfileOutput, ValidationIssue } from '.';
+import {
+  invalidWithErrors,
+  validWithWarnings,
+} from './test/validate-custom-matcher';
 import { formatIssues, getProfileOutput, validateMap } from './utils';
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace jest {
-    interface Matchers<R> {
-      toBeValidMap(
-        profileOutput: ProfileOutput,
-        warnings?: string[],
-        errors?: string[]
-      ): R;
-    }
-  }
-}
-
-expect.extend({
-  toBeValidMap(
-    map: MapASTNode,
-    profileOutput: ProfileOutput,
-    warnings?: string[],
-    errors?: string[]
-  ) {
-    const result = validateMap(profileOutput, map);
-
-    let message = '';
-    let pass = true;
-
-    const issues: { errors: ValidationIssue[]; warnings: ValidationIssue[] } = {
-      errors: [],
-      warnings: [],
-    };
-
-    if (!result.pass) {
-      issues.errors = result.errors;
-    }
-    if (result.warnings && result.warnings.length > 0) {
-      issues.warnings = result.warnings;
-    }
-
-    if (this.isNot) {
-      pass = false;
-
-      if (!errors) {
-        return {
-          pass: !pass,
-          message: () => 'Expected to fail, specify the errors',
-        };
-      }
-
-      if (result.pass) {
-        return {
-          pass: !pass,
-          message: () => 'Expected to fail, specified map is valid',
-        };
-      }
-
-      const err = formatIssues(issues.errors);
-      const warn = formatIssues(issues.warnings);
-
-      message = 'Expected to find errors:\n';
-
-      for (const error of errors) {
-        if (!err.includes(error)) {
-          if (!pass) {
-            pass = true;
-          }
-
-          message += `"${error}"\n`;
-        }
-      }
-
-      message += `in original errors:\n"${err}"\n`;
-
-      if (warnings && warnings.length > 0) {
-        message += '\nExpected to find warnings:\n';
-
-        for (const warning of warnings) {
-          if (!warn.includes(warning)) {
-            if (!pass) {
-              pass = true;
-            }
-
-            message += `"${warning}"\n`;
-          }
-        }
-
-        message += `in original warnings:\n"${warn}"\n`;
-      }
-    } else {
-      const warn = formatIssues(issues.warnings);
-      const err = formatIssues(issues.errors);
-
-      if (!result.pass && result.errors.length > 0) {
-        return {
-          pass: !pass,
-          message: () =>
-            `Expected to pass, specified map is invalid.\nErrors:\n${err}\nWarnings:\n${warn}\n`,
-        };
-      }
-
-      if (warnings && warnings.length > 0) {
-        message += 'Expected to find warnings:\n';
-
-        for (const warning of warnings) {
-          if (!warn.includes(warning)) {
-            if (pass) {
-              pass = false;
-            }
-
-            message += `"${warning}"\n`;
-          }
-        }
-
-        message += `in original warnings:\n"${warn}"\n`;
-      }
-    }
-
-    return {
-      pass,
-      message: (): string => message,
-    };
-  },
-});
-
-function validWithWarnings(
-  profile: ProfileDocumentNode,
-  maps: MapASTNode[],
-  ...warnings: string[][]
-): void {
-  const profileOutput = getProfileOutput(profile);
-
-  it('then validation will pass with warnings', () => {
-    maps.forEach((map, index) => {
-      expect(map).toBeValidMap(profileOutput, warnings[index]);
-    });
-  });
-}
-
-function invalidWithErrors(
-  profile: ProfileDocumentNode,
-  maps: MapASTNode[],
-  ...results: string[][]
-): void {
-  const profileOutput = getProfileOutput(profile);
-
-  it('then validation will fail with errors', () => {
-    let i = 0;
-    maps.forEach(map => {
-      expect(map).not.toBeValidMap(profileOutput, results[i + 1], results[i]);
-      i += 2;
-    });
-  });
-}
 
 function getIssues(profile: ProfileDocumentNode, maps: MapASTNode[]) {
   const profileOutput = getProfileOutput(profile);
